- Get rid of most Eclipse warning, although not yet with Java5   annotations. This will need to wait until we can get J2ME Polish to   postcompile it properly. - Fix the new KeyManager - contains(key) is wrong! Need to use   containsKey(key) instead on Hashtable. Damn, with Generics, this   could have been avoided. JUnit tests run through again. Yippeh.
git-svn-id: file:///srv/svn/openuat-filtered-without-thirdparty/trunk@727 01b9256e-9321-419f-88ee-03bd83f56c6c

+++ src/org/openuat/features/FPIntFFT.java
@@ -152,7 +152,7 @@ public class FPIntFFT {
   Scaling ensures that result remains 16-bit. */
 	private final static short FIX_MPY(short a, short b) {
 		/* shift right one less bit (i.e. 15-1) */
-		int c = ((int)a * (int)b) >> 14;
+		int c = (a * b) >> 14;
 		/* last bit shifted out = rounding-bit */
 		b = (short) (c & 0x01);
 		/* last shift + rounding bit */
